feat: parallel batch [T20260508-1]
Tasks:
- T20260508-1: Break down crates/orbit-core/src/runtime/v2_host.rs into focused submodules
